Attach users photo_consents for event
Previously, we checked photo_consent status for an event in the frontend
by comparing the users registered photo consents with the event
start_time.

This among other things, led to a bug where the user does not have any
photo_consents for an event if the event is in the next semester, and
the user cannot attend the event if `use_consent=True`. This fixes that
by attaching and creating photo consents for that specific event.